requirements.txt
# 数据获取和处理
akshare>=1.12.0
pandas>=1.5.0
numpy>=1.24.0

# 数据分析和计算
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
mplfinance>=0.12.0

# 回测框架
backtrader>=1.9.78
zipline>=1.4.1

# 配置和工具
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.25.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 其他工具
requests>=2.31.0
lxml>=4.9.0
beautifulsoup4>=4.12.0 